feat: add option to open the full commit message in a separate tab
Add a new configuration option: "showEditor" that displays
the commit message in a separate tab as a text document the user can edit.
If autoCommit is set to true, saving the file will automatically commit the
message. If autoCommit is set to false, it will paste the contents
into the scm as usual and close the tab.